+++ CHANGELOG.md
@@ -7,3 +7,11 @@
 
 ### Third Party Javascript Copy Added
 - Third party script files copy content
+
+## v2.0.0-beta (June 14, 2018)
+
+### Major Updates
+- Updated Webpack version to 4.x.x
+- changed extract-text plugin to min css extract plugin
+- added Webpack command module
+- Changed the engine configs